requirements.txt
pytesseract>=0.3.9
Pillow>=5.4.1
pdf2image>=1.16.0